credit-card-fraud-detection/requirements.txt
streamlit==1.13.0
joblib==1.2.0
numpy==1.23.4
pandas==1.5.1
matplotlib==3.6.0
seaborn==0.12.1
Pillow==9.2.0
scikit-learn==1.1.2
